% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_resolution.R
\name{beta_coal_block_size_alpha_mle}
\alias{beta_coal_block_size_alpha_mle}
\title{Estimate of alpha parameter of a multifurcating tree under
the Beta-coalescent model using only topology information}
\usage{
beta_coal_block_size_alpha_mle(tree)
}
\arguments{
\item{tree}{A \code{phylo} object}
}
\value{
The block-size MLE for estimating alpha
}
\description{
Estimate of alpha parameter of a multifurcating tree under
the Beta-coalescent model using only topology information
}
