% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate-plot.R
\name{BNPR}
\alias{BNPR}
\alias{BNPR_PS}
\title{Bayesian nonparametric phylodynamic reconstruction.}
\usage{
BNPR(data, lengthout = 300, pref = FALSE, prec_alpha = 0.01,
  prec_beta = 0.01, beta1_prec = 0.001, log_zero = -100)

BNPR_PS(data, lengthout = 300, prec_alpha = 0.01, prec_beta = 0.01,
  beta1_prec = 0.001, log_zero = -100)
}
\arguments{
\item{data}{\code{phylo} object or list containing vectors of coalescent
times \code{coal_times}, sampling times \code{s_times}, and number sampled
per sampling time \code{n_sampled}.}

\item{lengthout}{numeric specifying number of grid points - 1.}

\item{pref}{logical. Should the preferential sampling model be used?}

\item{prec_alpha,prec_beta}{numerics specifying gamma prior for precision
\eqn{\tau}.}

\item{beta1_prec}{numeric specifying precision for normal prior on
\eqn{\beta_1}.}

\item{log_zero}{numeric specifying approximate log of zero.}
}
\value{
Phylodynamic reconstruction of (log) effective population size at the
  grid points.
}
\description{
Bayesian nonparametric phylodynamic reconstruction.
}
\section{Functions}{
\itemize{
\item \code{BNPR_PS}: Uses preferential sampling model.
}}
\examples{
data("NY_flu")
res = BNPR(NY_flu)
plot_BNPR(res)
}

