% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_resolution.R
\name{BNPR_Lambda}
\alias{BNPR_Lambda}
\title{Bayesian nonparametric phylodynamic reconstruction of multifurcating tree 
under the Lambda-coalescent}
\usage{
BNPR_Lambda(
  data,
  alpha_multif,
  lambda_meas = NULL,
  lengthout = 100,
  pref = FALSE,
  prec_alpha = 0.01,
  prec_beta = 0.01,
  beta1_prec = 0.001,
  fns = NULL,
  log_fns = TRUE,
  simplify = TRUE,
  derivative = FALSE,
  forward = TRUE,
  grid = NULL
)
}
\arguments{
\item{data}{\code{phylo} object or list containing vectors of coalescent 
times \code{coal_times}, sampling times \code{samp_times}, block sizes
\code{block_sizes}, and number sampled per sampling time \code{n_sampled}.}

\item{alpha_multif}{number between 0 and 2 giving the parameter for the Beta coalescent}

\item{lambda_meas}{A probability density function on the unit interval, or a named list 
of \code{values} and \code{probs} representing a probaiblity mass function on the 
unit interval. Not used if \code{alpha_multif} is specified.}

\item{lengthout}{numeric specifying number of grid points.}

\item{pref}{logical. Should the preferential sampling model be used? Default is \code{FALSE}.}

\item{prec_alpha, prec_beta}{numerics specifying gamma prior for precision \eqn{\kappa}.}

\item{beta1_prec}{numeric specifying precision for normal prior on \eqn{\beta_1}.}

\item{fns}{list containing functions of covariates.}

\item{log_fns}{logical whether or not to to apply a log-transformation to
the output of the functions in \code{fns}.}

\item{simplify}{logical whether to fully bucket all Poisson points.}

\item{derivative}{logical whether to calculate estimates of the 
log-derivative.}

\item{forward}{logical whether to use the finite difference approximations of
the log-derivative as a forward or backward derivative.}

\item{grid}{it gives the opportunity to evaluate BNPR on a prespecified grid}
}
\value{
Phylodynamic reconstruction of effective population size at grid 
  points. \code{result} contains the INLA output, \code{data} contains the 
  information passed to INLA, \code{grid} contains the grid end points, 
  \code{x} contains the grid point centers, \code{effpop} contains a vector 
  of the posterior median effective population size estimates, 
  \code{effpop025} and \code{effpop975} contain the 2.5th and 97.5th 
  posterior percentiles, \code{summary} contains a data.frame of the 
  estimates, and \code{derivative} (if \code{derivative = TRUE}) contains a
  data.frame summarizing the log-derivative.
}
\description{
Bayesian nonparametric phylodynamic reconstruction of multifurcating tree 
under the Lambda-coalescent
}
