% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{BNPR_multiple}
\alias{BNPR_multiple}
\title{Bayesian nonparametric phylodynamic reconstruction from multiple correlated trees}
\usage{
BNPR_multiple(
  data,
  lengthout = 100,
  prec_alpha = 0.01,
  prec_beta = 0.01,
  zero_dates = NA
)
}
\arguments{
\item{data}{\code{multiPhylo} multiPhylo object}

\item{lengthout}{numeric specifying number of grid points.}

\item{prec_alpha, prec_beta}{numerics specifying gamma prior for precision 
\eqn{\tau}.}

\item{zero_dates, }{a vector with actual dates of the most recent sample per phylogeny}
}
\value{
Phylodynamic reconstruction of effective population size at grid 
  points. \code{data} contains the information passed to INLA, \code{grid} contains the grid end points, 
  \code{x} contains the grid point centers, \code{effpop} contains a vector 
  of the posterior median effective population size estimates, 
  \code{effpop025} and \code{effpop975} contain the 2.5th and 97.5th 
  posterior percentiles, \code{summary} contains a data.frame of the 
  estimates, and \code{derivative} (if \code{derivative = TRUE}) contains a
  data.frame summarizing the log-derivative.
}
\description{
Bayesian nonparametric phylodynamic reconstruction from multiple correlated trees
}
\examples{
data("NY_flu")
res = BNPR(NY_flu)
plot_BNPR(res)
}
