% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{boombust_traj}
\alias{boombust_traj}
\title{Boom and bust trajectory}
\usage{
boombust_traj(t, bust = 1, scale = 1000)
}
\arguments{
\item{t}{numeric vector of times at which to evaluate the trajectory 
function.}

\item{bust}{numeric time when bust begins.}

\item{scale}{numeric maximum that the trajectory attains when the bust
begins.}
}
\value{
A vector containing the trajectory value corresponding to every 
  element in \code{t}.
}
\description{
Boom and bust trajectory
}
\examples{
boombust_traj(0:10)
}

