% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Bottle_20c}
\alias{Bottle_20c}
\title{Bottle_20c}
\format{An object of class \code{multiPhylo} of length 1435.}
\usage{
Bottle_20c
}
\description{
This dataset is a \code{multiPhylo} object.
}
\keyword{datasets}

