% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{bottleneck_traj}
\alias{bottleneck_traj}
\title{Bottleneck trajectory}
\usage{
bottleneck_traj(t)
}
\arguments{
\item{t}{numeric vector of times at which to evaluate the trajectory 
function.}
}
\value{
A vector containing the trajectory value corresponding to every 
  element in \code{t}.
}
\description{
Bottleneck trajectory
}
\examples{
bottleneck_traj(seq(0,2,by=0.1))
}

