% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounded_coal_functions.R
\name{bounded_skyline_ascent}
\alias{bounded_skyline_ascent}
\title{MLE estimation of Ne from bounded coalescent.}
\usage{
bounded_skyline_ascent(data, bound = 1)
}
\arguments{
\item{data}{\code{phylo} object or list containing vectors of coalescent 
times \code{coal_times}, sampling times \code{samp_times}, and number 
sampled per sampling time \code{n_sampled}.}

\item{bound}{provide uper bound on TMRCA}
}
\value{
Phylodynamic reconstruction of effective population size at grid 
  points.

Phylodynamic reconstruction of effective population size at grid 
  points.
}
\description{
MLE estimation of Ne from bounded coalescent.
}
\examples{
data<-coalsim_bounded(c(0),c(10),constant,bound=0.5)
bounded_skyline(data, lengthout=10, bound=0.5)
MLE estimation of Ne from bounded coalescent.

data<-coalsim_bounded(c(0),c(10),constant,bound=0.5)
bounded_skyline_ascent(data, lengthout=10, bound=0.5)
}
