% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_check.R
\name{coal_check}
\alias{coal_check}
\title{Predictive check for coalescent model}
\usage{
coal_check(coal_times, samp_times, n_sampled, logpop, grid)
}
\arguments{
\item{coal_times}{numeric vector of coalescent times.}

\item{samp_times}{numeric vector of sampling times.}

\item{n_sampled}{integer vector of lineages sampled at each sampling time.}

\item{logpop}{numeric vector of log-effective population sizes.}

\item{grid}{numeric grid of time points.}
}
\value{
named numeric vector: "obs" contains the KS discrepancy for the
  supplied coalescent times, "rep" contains å KS discrepancy for the
  replicated coalescent times.
}
\description{
Predictive check for coalescent model
}
