% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coalsieve.R
\name{coalgen_thinning_hetero}
\alias{coalgen_thinning_hetero}
\title{Generate inhomogeneous coalescent with heterochronous samples.}
\usage{
coalgen_thinning_hetero(sample, traj_inv, upper, ...)
}
\arguments{
\item{sample}{a matrix with 2 columns. The first column contains the number
of samples collected at the time defined in the second column.}

\item{traj_inv}{a function returning the one over the effective population
size.}

\item{upper}{numeric. The maximum of the traj_inv function on the relevant
interval.}

\item{...}{additional parameters to pass to the traj_inv function.}
}
\value{
A list containing coalescent times \code{coal_times}, intercoalescent
  times \code{intercoal_times}, and number of active lineages between
  coalescent times \code{lineages}.
}
\description{
Generate inhomogeneous coalescent with heterochronous samples.
}
\examples{
coalgen_thinning_hetero(cbind(c(10,3), c(0,5)), unif_traj_inv, 0.01)
}

