% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounded_coal_functions.R
\name{coalsim_bounded}
\alias{coalsim_bounded}
\title{Simulate from inhomogeneous, heterochronous bounded coalescent}
\usage{
coalsim_bounded(
  samp_times,
  n_sampled,
  traj = constant,
  bound = 1,
  val_upper = 10,
  ...
)
}
\arguments{
\item{samp_times}{numeric vector of sampling times.}

\item{n_sampled}{numeric vector of samples taken per sampling time.}

\item{traj}{function that returns effective population size at time t.}

\item{bound}{upper bound on the TMRCA}

\item{val_upper}{upper value for root finding in transformation method}
}
\value{
A list containing vectors of coalescent times \code{coal_times}, 
  intercoalescent times \code{intercoal_times}, and number of active lineages
  \code{lineages}, as well as passing along \code{samp_times} and
  \code{n_sampled}.
}
\description{
Simulate from inhomogeneous, heterochronous bounded coalescent
}
\examples{
coalsim_bounded(0:2, 3:1, unif_traj, bound=1)
}
