% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coal_lik_BSMC.R
\name{create_F}
\alias{create_F}
\title{Generates a matrix with ranked tree shape information}
\usage{
create_F(tree)
}
\arguments{
\item{tree}{a phylo object}
}
\value{
An F matrix containing ranked tree shape information
}
\description{
Generates a matrix with ranked tree shape information
}
\examples{
create_F(ape::rcoal(4))
}

