% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{cyclic_traj}
\alias{cyclic_traj}
\title{Cyclical trajectory}
\usage{
cyclic_traj(t)
}
\arguments{
\item{t}{numeric vector of times at which to evaluate the trajectory 
function.}
}
\value{
A vector containing the trajectory value corresponding to every 
  element in \code{t}.
}
\description{
Cyclical trajectory
}
\examples{
cyclic_traj(0:20)
}

