% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_resolution.R
\name{dev_new}
\alias{dev_new}
\title{Computes the deviance of a reconstructed EPS trajectory given a reference trajectory}
\usage{
dev_new(INLA_out, traj, hilim = Inf, lolim = 0, yhilim = Inf, ylolim = 0, ...)
}
\arguments{
\item{INLA_out}{output directly from \code{BNPR_Lambda} or \code{posterior_beta_coal_MCMC_w_sHMC},
or the \code{BNPR} argument from the output of \code{posterior_beta_coal_hybrid}}

\item{traj}{the reference trajectory to compare to}
}
\value{
List of total deviance \code{total} or averaged deviance by number of grid points \code{avg}.
}
\description{
Computes the deviance of a reconstructed EPS trajectory given a reference trajectory
}
