% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_distance.R
\name{dist_Amat}
\alias{dist_Amat}
\title{Distance between two A-matrices}
\usage{
dist_Amat(A1, A2, norm = "L1")
}
\arguments{
\item{A1}{The first A-matrix.}

\item{A2}{The second A-matrix.}

\item{norm}{The norm for calculating distance, either "L1" or "L2".}
}
\value{
The distance between two A-matrices.
}
\description{
This function calculates the distance between two input A-matrices.
}
\examples{
t1 <- ape::rcoal(5)
t2 <- ape::rcoal(5)
A1 <- gen_Amat(t1)
A2 <- gen_Amat(t2)
d <- dist_Amat(A1, A2, "L1")
}
