% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_distance.R
\name{dist_Fmat}
\alias{dist_Fmat}
\title{Distance between two F-matrices}
\usage{
dist_Fmat(F1, F2, norm = "L1")
}
\arguments{
\item{F1}{The first F-matrix.}

\item{F2}{The second F-matrix.}

\item{norm}{The norm for calculating distance, either "L1" or "L2".}
}
\value{
The distance between two F-matrices.
}
\description{
This function calculates the distance between two input F-matrices.
}
\examples{
t1 <- ape::rcoal(5)
t2 <- ape::rcoal(5)
F1 <- gen_Fmat(t1)
F2 <- gen_Fmat(t2)
d <- dist_Fmat(F1, F2, "L2")
}
