% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_distance.R
\name{dist_Fmat_weighted}
\alias{dist_Fmat_weighted}
\title{Distance between two F-matrices weighted by branch lengths}
\usage{
dist_Fmat_weighted(F1, F2, W1, W2, norm = "L1")
}
\arguments{
\item{F1}{The first F-matrix.}

\item{F2}{The second F-matrix.}

\item{norm}{The norm for calculating distance, either "L1" or "L2".}
}
\value{
The distance between two F-matrices weighted by branch lengths.
}
\description{
This function calculates the distance between two input F-matrices weighted by branch lengths.
}
\examples{
t1 <- ape::rcoal(5)
t2 <- ape::rcoal(5)
F1 <- gen_Fmat(t1)
F2 <- gen_Fmat(t2)
W1 <- gen_Wmat(t1)
W2 <- gen_Wmat(t2)
d <- dist_Fmat_weighted(F1, F2, W1, W2, "L1")
}
