% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_distance.R
\name{dist_label_tree}
\alias{dist_label_tree}
\title{Distance between labeled trees}
\usage{
dist_label_tree(t1, t2, alpha = 0.5, weighted = F, norm = "L1")
}
\arguments{
\item{t1}{A \code{phylo} object.}

\item{t2}{A \code{phylo} object, with the same number of tips as t1.}

\item{alpha}{A number between 0 and 1. The weight we put on F-matrix. 1-alpha is the weight we put on A-matrix.}

\item{norm}{The norm for calculating distance, either "L1" or "L2".}

\item{weight}{A logical variable indicating if branch length information is considered.}
}
\value{
The distance between two labeled trees.
}
\description{
This function generates the distance between two labeled trees with the same number of tips. 
It is calculated as a weighted average of F-matrix norm and A-matrix norm.
}
\examples{
t1 <- ape::rcoal(5)
t2 <- ape::rcoal(5)
d <- dist_label_tree(t1, t2)
}
