% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance.R
\name{dist_pairwise}
\alias{dist_pairwise}
\title{Distance between two ranked tree shapes or ranked genealogies}
\usage{
dist_pairwise(tr.1, tr.2, dist.method = "l1", weighted = FALSE)
}
\arguments{
\item{tr.1}{An object of class \code{phylo}. Can be either isochronous or heterochronous.}

\item{tr.2}{An object of class \code{phylo}. Can be either isochronous or heterochronous.}

\item{dist.method}{"l1" or "l2".}

\item{weighted}{A logical value; if \code{TRUE}, weighted distance is computed.}
}
\value{
Distance between the input trees \code{tr.1} and \code{tr.2}.
}
\description{
Distance between two ranked tree shapes or ranked genealogies
}
\details{
The input trees, tr.1 and tr.2 must have the same number of taxa.
The number of sampling events, however, can differ between two trees.

\code{dist.method="l1"} option computes Manhattan distance between
F-matrices (flattened as vectors) of tr.1 and tr.2.

\code{dist.method="l2"} option computes Frobenius distance between
F-matrices (flattened as vectors) of tr.1 and tr.2.

If \code{weighted=TRUE}, the F-matrix elements are weighted by
the corresponding time intervals.

If \code{weighted=FALSE}, the F-matrix elements are not weighted by
the corresponding time intervals, i.e., distance is computed using
ranked tree shapes only.
}
\examples{
# Generate a sample tree
set.seed(1); tr.1 <- ape::rcoal(5)
set.seed(2); tr.2 <- ape::rcoal(5)
dist_pairwise(tr.1, tr.2, dist.method='l1', weighted=TRUE)

}
\references{
Kim J, Rosenberg NA, Palacios JA, 2019.
\emph{A Metric Space of Ranked Tree Shapes and Ranked Genealogies}.
}
\author{
Jaehee Kim, Julia Palacios
}
