% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_resolution.R
\name{envelope_new}
\alias{envelope_new}
\title{Computes the coverage of the 95% credible interval of a reconstructed EPS 
trajectory given a reference trajectory}
\usage{
envelope_new(
  INLA_out,
  traj,
  hilim = Inf,
  lolim = 0,
  yhilim = Inf,
  ylolim = 0,
  ...
)
}
\arguments{
\item{INLA_out}{output directly from \code{BNPR_Lambda} or \code{posterior_beta_coal_MCMC_w_sHMC},
or the \code{BNPR} argument from the output of \code{posterior_beta_coal_hybrid}}

\item{traj}{the reference trajectory to compare to}
}
\value{
List of number of grid points covered \code{total} or percentage covered \code{avg}.
}
\description{
Computes the coverage of the 95% credible interval of a reconstructed EPS 
trajectory given a reference trajectory
}
