% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SLT.R
\name{estimate_mu}
\alias{estimate_mu}
\title{Estimation of mutation rates via Method-of-Moments}
\usage{
estimate_mu(data, t)
}
\arguments{
\item{data}{rows correspond to cells and columns to S target sites}

\item{t}{time to the most recent common ancestor
It returns a vector where the first S values are individual target mutations, followed by 
S(S-1)/2 off-diagonal(by row)}
}
\description{
This function estimates the rates of the CTMC
}
