% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastaToxml_beast2.R
\name{fastaToxml_beast2}
\alias{fastaToxml_beast2}
\title{Generate xml files from fasta files}
\usage{
fastaToxml_beast2(inputfasta, outputxml, xml_template, clock_rate)
}
\arguments{
\item{inputfasta}{The path to the input fasta file}

\item{outputxml}{The path to the output xml file}

\item{xml_template}{The path to the xml template}

\item{clock_rate}{The clock rate used in BEAUti}
}
\value{
An xml file
}
\description{
Given a fasta file from seqgen(), this function outputs an xml file as if we run BEAUti
}
\examples{
fastaToxml_beast2("/path/inputfasta.fasta", "/path/outputxml.xml", "/path/xml_template.txt", clock_rate = 2) 
}
