% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coal_lik_BSMC.R
\name{find_info2}
\alias{find_info2}
\title{Extracts all sufficient statistics for inferring Ne from sequential local genealogies}
\usage{
find_info2(MyTree, D, sim, tol, cor = 1)
}
\arguments{
\item{MyTree}{\code{multiPhylo} object containing coalescent trees}

\item{D}{a matrix of coalescent times}

\item{sim}{integer specifying number of trees to read}

\item{tol}{tolerance for detecting difference between two numeric values}

\item{cor}{(internal)}
}
\value{
A list of sufficient statistics
}
\description{
Extracts all sufficient statistics for inferring Ne from sequential local genealogies
}

