% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_distance.R
\name{gen_Amat}
\alias{gen_Amat}
\title{Generation of A-matrix}
\usage{
gen_Amat(t)
}
\arguments{
\item{t}{A \code{phylo} object.}
}
\value{
A-matrix of the input tree.
}
\description{
This function generates the A-matrix that encodes the labeling information of an input tree.
}
\examples{
t <- ape::rcoal(5)
Amat <- gen_Amat(t)
}
