% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance.R
\name{gen_Fmat}
\alias{gen_Fmat}
\title{Unique encoding of ranked tree shapes}
\usage{
gen_Fmat(tr, tol = 13)
}
\arguments{
\item{tr}{An object of class \code{phylo}. Can be either isochronous or heterochronous.}

\item{tol}{Numerical tolerance.}
}
\value{
F-matrix of the input tree.
}
\description{
This function generates a F-matrix, a matrix that encodes
a ranked tree shape of an input tree.
Works for both isochronous and heterochronous trees.
}
\details{
The time starts from zero at the first sampling event \eqn{u_{n+m-1}}.
An F-matrix that encodes the ranked tree shape of the input \code{tr} is
a lower triangular matrix of integers with elements \eqn{F_{i,j}=0} for all
\eqn{i < j}, and for \eqn{1 \le j \le i}, \eqn{F_{i,j}} is the number of
extant lineages in \eqn{(u_{j+1}, u_{j})} that do not bifurcate or become
extinct during the entire time internval \eqn{(u_{i+1}, u_{j})}.

\code{tol} is introduced to fix the issue that nodes sampled at the same time
are treated as samples from different sampling times due to numerical error.
Here, if the samples are sampled at times within \code{tol}, they're treated
as sampled at the same time.
}
\examples{
# Generate a sample tree
tr <- ape::rcoal(5)
Fmat <- gen_Fmat(tr)

}
\references{
Kim J, Rosenberg NA, Palacios JA, 2019.
\emph{A Metric Space of Ranked Tree Shapes and Ranked Genealogies}.
}
\author{
Jaehee Kim, Julia Palacios
}
