% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_distance.R
\name{gen_Wmat}
\alias{gen_Wmat}
\title{Generation of W-matrix}
\usage{
gen_Wmat(t)
}
\arguments{
\item{t}{A \code{phylo} object.}
}
\value{
W-matrix of the input tree.
}
\description{
This function generates the W-matrix that encodes the branch length information of an input tree.
}
\examples{
t <- ape::rcoal(5)
Wmat <- gen_Wmat(t)
}
