% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newick_Files.R
\name{generate_newick}
\alias{generate_newick}
\title{Returns a phylo object from the argumentes generated with coalsim}
\usage{
generate_newick(args)
}
\arguments{
\item{args}{is a list containing vectors of coalescent times \code{coal_times}, sampling times \code{samp_times}, and number 
sampled per sampling time \code{n_sampled}, etc. This list is the output of coalsim}
}
\value{
A list with two elements \code{newikck} contains the tree in phylo format, \code{lables} a vector with tip labels
}
\description{
Returns a phylo object from the argumentes generated with coalsim
}
\examples{
constant<-function(x){return (rep(1,length(x)))}
simulation1<-coalsim(0,10,constant)
tree1<-generate_newick(simulation1)
plot(tree1$newick)
}
