% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_resolution.R
\name{generate_newick_multif}
\alias{generate_newick_multif}
\title{Returns a phylo object from the argumentes generated with beta_coalsim}
\usage{
generate_newick_multif(args)
}
\arguments{
\item{args}{is a list containing vectors of coalescent times \code{coal_times}, sampling times \code{samp_times}, 
block sizes \code{block_sizes}, and number sampled per sampling time \code{n_sampled}, etc. This list is the output of coalsim}
}
\value{
A list with two elements \code{newikck} contains the tree in phylo format, \code{lables} a vector with tip labels
}
\description{
Returns a phylo object from the argumentes generated with beta_coalsim
}
\examples{
simulation1<-beta_coalsim(samp_times=c(0,1), n_sampled = c(25,25), alpha=1.5, traj=exp_traj)
tree1<- generate_newick_multif(simulation1)$newick
plot(tree1$newick)
}
