% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tajima_inference.R
\name{graph_gene_tree2}
\alias{graph_gene_tree2}
\title{Plot the perfect phylogeny along with a vector of mutations. Vignette Tajima_prepare&plotdata provides details on how to use it}
\usage{
graph_gene_tree2(suff, mutations)
}
\arguments{
\item{suff}{the nodes object of a perfect phylogeny}

\item{mutations}{list of mutations to add in the edges for plotting}
}
\description{
Plot the perfect phylogeny along with a vector of mutations. Vignette Tajima_prepare&plotdata provides details on how to use it
}
