% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tajima_inference.R
\name{graph_gene_tree_mut}
\alias{graph_gene_tree_mut}
\title{Plot perfect phylogeny}
\usage{
graph_gene_tree_mut(suff, mutations)
}
\arguments{
\item{suff}{is the output of tajima_perfect phylogeny}

\item{mutations}{is a vector recording the mutation on each edge of the perfect phylogeny}
}
\value{
plot automotically the perfect phylogeny
}
\description{
Plot perfect phylogeny
}
