% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist2heat.R
\name{hist2heat}
\alias{hist2heat}
\title{Plot heatmap of a histogram}
\usage{
hist2heat(hist, y, wd)
}
\arguments{
\item{hist}{\code{histogram} object to be displayed.}

\item{y}{numeric y-coordinate to display heatmap.}

\item{wd}{numeric width of heatmap in y-units.}
}
\description{
Plot heatmap of a histogram
}

