% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SLT.R
\name{inference_all}
\alias{inference_all}
\title{MLE estimation of topology and coalescent times}
\usage{
inference_all(Dlist, t0, thetas, M = 20, mu_list = NULL)
}
\arguments{
\item{Dlist}{a list  per barcode, of alleles. Each column is a target site}

\item{t0}{time to the most recent common ancestor}

\item{thetas}{are the mutation rate parameters per barcode}

\item{M}{is the number of alleles}
}
\value{
a genealogy

#Dlist of data (fulle alleles) matrices, corresponding to intBCs -- assume all lists the same
Uses the input M parameter (assumed to be the same for every intBC) if provided; otherwise
uses the counts of number of unique alleles in the dataset as M (max across intBCs)
This function does the full inference procedure:
.....1) Infers rates for each intBC (NOT assumed to be the same), if thetas is not input
.....2) Gets MLE upgma tree + initial branchlengths
.....3) Gradient ascent for branch lengths
}
\description{
MLE estimation of topology and coalescent times
}
