% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SLT.R
\name{inference_times}
\alias{inference_times}
\title{MLE estimation of coalescent times given tree and data}
\usage{
inference_times(Dlist, thetas, M = 20, tree, mu_list = NULL)
}
\arguments{
\item{Dlist}{a list  per barcode, of alleles. Each column is a target site}

\item{thetas}{are the mutation rate parameters per barcode}

\item{tree}{is the known ape genealogy}
}
\value{
a vector of inter-coalescent times
}
\description{
MLE estimation of coalescent times given tree and data
}
