% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phyloinfer.R
\name{initial_MCMC}
\alias{initial_MCMC}
\title{BESTT MCMC initialization}
\usage{
initial_MCMC(
  initial,
  ngrid = 50,
  mu = 40,
  n = 10,
  alpha = 0.1,
  fact = 10,
  Nate = 2000
)
}
\arguments{
\item{initial}{output of initial_tajima}

\item{ngrid}{#grid points}

\item{mu}{Mutation rate}

\item{n}{sample size}

\item{alpha}{hyperparameter}

\item{fact}{scaling factor}

\item{Nate}{#initialization iterations}
}
\value{
A list with initial values.
}
\description{
BESTT MCMC initialization
}
