% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tajima_inference.R
\name{initial_MCMC_het}
\alias{initial_MCMC_het}
\title{Initial MCMC fuction}
\usage{
initial_MCMC_het(
  initial,
  ngrid = 50,
  mu,
  n,
  alpha = 0.1,
  fact,
  Nate,
  times,
  samp_times
)
}
\arguments{
\item{initial}{output of initial_tajima_het}

\item{ngrid}{grid size}

\item{mu}{mutation rate}

\item{n}{vector sample size}

\item{alpha}{precision parameter cov matrix Gaussian process}

\item{fact}{scaling factor (if time or N_e(t) is rescaled)}

\item{Nate}{number of initialization steps}

\item{times}{input initial$times}

\item{samp_times}{vector sampling times}
}
\value{
initialization MCMC
}
\description{
Initial MCMC fuction
}
