% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phyloinfer.R
\name{initial_tajima}
\alias{initial_tajima}
\title{BESTT initialization}
\usage{
initial_tajima(data1, name = "newSim10Bottle", mu, npoints = 49, fact = 1)
}
\arguments{
\item{name}{The name of the fasta file to be stored in your working directory}

\item{mu}{Mutation rate}

\item{npoints}{Number of breakpoints of a regular grid to initialize Ne}

\item{fact}{scaling factor}

\item{data}{A vector of mutations}
}
\value{
A list with initial values. theta is log Ne
}
\description{
BESTT initialization
}
