% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tajima_inference.R
\name{initial_tajima_het}
\alias{initial_tajima_het}
\title{Sample heterochronous topology.}
\usage{
initial_tajima_het(
  data1,
  name = "newSim10Bottle",
  mu,
  npoints = 49,
  fact = 1,
  alpha = 0.2,
  samp_times,
  n,
  input_tree,
  name_samp
)
}
\arguments{
\item{data1}{sequence data as sampled by simulate_data_het}

\item{name}{name of the file that converts sequence data into nucleotide basis}

\item{mu}{mutation rate}

\item{npoints}{#grid points to approximate N_e(t)}

\item{fact}{scaling factor (if time or N_e(t) is rescaled)}

\item{alpha}{precision parameter cov matrix Gaussian process}

\item{samp_times}{vector sampling times}

\item{n}{vector sample size}

\item{input_tree}{input tree}

\item{name_samp}{two columns in which one is the sampling times paired with an arbitrary sequence label.}
}
\value{
initialization list
}
\description{
Sample heterochronous topology.
}
