% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phyloinfer.R
\name{initialize_params}
\alias{initialize_params}
\title{Set fixed parameters for prior, likelihood, and posterior calculations.}
\usage{
initialize_params(samp_times, n_sampled, coal_times, grid,
  prec_alpha = 0.01, prec_beta = 0.01, betas_prec = diag(c(0.01,
  0.01)), covar_vals = NULL)
}
\arguments{
\item{samp_times}{vector of sampling times.}

\item{n_sampled}{integer vector of number of tips sampled at each sampling time.}

\item{coal_times}{vector of coalescent times.}

\item{grid}{vector of grid end points.}

\item{prec_alpha}{numeric alpha hyperparameter for precision parameter (default 0.01).}

\item{prec_beta}{numeric beta hyperparameter for precision parameter (default 0.01).}

\item{betas_prec}{matrix precision matrix for beta parameters.}

\item{covar_vals}{matrix of covariate values at each of the grid midpoints.}
}
\value{
list containing all the parameters necessary to use 
logprior_[], loglik_[] and logposterior_[] functions.
}
\description{
Set fixed parameters for prior, likelihood, and posterior calculations.
}
