% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{inla.models}
\alias{inla.models}
\title{INLA wrapper}
\usage{
inla.models(...)
}
\arguments{
\item{...}{passes all arguments to \code{INLA::inla.models}.}
}
\description{
Wrapper for \code{INLA::inla.models} to work around a potential namespace bug
in \code{INLA::inla}.
}

