% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function countingv8.R
\name{king_initial_sampnodes}
\alias{king_initial_sampnodes}
\title{Define a list of nodes to sample from}
\usage{
king_initial_sampnodes(nodes)
}
\arguments{
\item{nodes}{Kingman perfect phylogenty}
}
\value{
active sampling nodes
}
\description{
Define a list of nodes to sample from
}
