% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function countingv8.R
\name{kingman_perphylo}
\alias{kingman_perphylo}
\title{Compute Kingman perfect phylogeny}
\usage{
kingman_perphylo(nodes)
}
\arguments{
\item{nodes}{it is the matrix description of the perfect phylogeny as describe by $nodes of the output of sufficent_stats}
}
\value{
PP
}
\description{
Compute Kingman perfect phylogeny
}
