% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function countingv8.R
\name{label_king}
\alias{label_king}
\title{Add labels to the sequences}
\usage{
label_king(nodes, sampnodes)
}
\arguments{
\item{nodes}{Kingman perfect phylogeny}

\item{sampnodes}{Kingman active sampling nodes}
}
\value{
labeled information
}
\description{
Add labels to the sequences
}
