% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{log_samp_mat}
\alias{log_samp_mat}
\title{Calculate posterior sampling intensities}
\usage{
log_samp_mat(popTbl, betaTbl, covariates = NULL, powBetaTbl = NULL,
  pow_covariates = NULL)
}
\arguments{
\item{popTbl}{table of log-effective population sizes.}

\item{betaTbl}{table of log-linear sampling model coefficients.}

\item{covariates}{list of covariates}

\item{powBetaTbl}{table of log-linear sampling model coefficients for terms 
crossed with the log-effective population size.}

\item{pow_covariates}{list of covariates to be crossed with the log-effective
population size.}
}
\value{
a matrix of posterior sampling intensities.
}
\description{
Calculate posterior sampling intensities
}
\examples{
2 + 2
}

