% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{logistic_traj}
\alias{logistic_traj}
\title{Seasonal logistic trajectory}
\usage{
logistic_traj(t, offset = 0, a = 2, period = 12, min = 10, max = 100)
}
\arguments{
\item{t}{numeric vector of times at which to evaluate the trajectory 
function.}

\item{offset}{numeric. By default \code{t=0} is a local maximum. Offset is
added to \code{t} to allow \code{t=0} to represent a different point in the
cycle.}

\item{a}{numeric slopedness parameter.}
}
\value{
A vector containing a value of \code{level} for every element in
  \code{t}.
}
\description{
Seasonal logistic trajectory
}
\examples{
unif_traj(0:12)
}

