% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phyloinfer.R
\name{loglik_ESS}
\alias{loglik_ESS}
\title{Compute the log-likelihood using the set-up for ESS (sampling conditional)}
\usage{
loglik_ESS(f, params)
}
\arguments{
\item{f}{vector of log effective population sizes at grid midpoints.}

\item{params}{output from initialize_parameters().}
}
\value{
numeric log-likelihood to agree with the output of mcmc_sampling(alg='ESS').
}
\description{
Compute the log-likelihood using the set-up for ESS (sampling conditional)
}
