% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phyloinfer.R
\name{loglik_ESS_pref}
\alias{loglik_ESS_pref}
\title{Compute the log-likelihood using the set-up for ESS (sampling aware)}
\usage{
loglik_ESS_pref(f, betas, params)
}
\arguments{
\item{f}{vector of log effective population sizes at grid midpoints.}

\item{betas}{vector of beta coefficients for log-linear sampling model.}

\item{params}{output from initialize_parameters().}
}
\value{
numeric log-likelihood to agree with the output of mcmc_sampling(alg='ESS', samp_alg='MH').
}
\description{
Compute the log-likelihood using the set-up for ESS (sampling aware)
}
