% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phyloinfer.R
\name{logprior_ESS}
\alias{logprior_ESS}
\title{Compute the log-prior using the set-up for ESS (sampling conditional)}
\usage{
logprior_ESS(f, kappa, params)
}
\arguments{
\item{f}{vector of log effective population sizes at grid midpoints.}

\item{kappa}{numeric precision parameter.}

\item{params}{output from initialize_parameters().}
}
\value{
numeric log-prior to agree with the output of mcmc_sampling(alg='ESS').
}
\description{
Compute the log-prior using the set-up for ESS (sampling conditional)
}
