% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phyloinfer.R
\name{logprior_ESS_pref}
\alias{logprior_ESS_pref}
\title{Compute the log-prior using the set-up for ESS (sampling aware)}
\usage{
logprior_ESS_pref(f, kappa, betas, params)
}
\arguments{
\item{f}{vector of log effective population sizes at grid midpoints.}

\item{kappa}{numeric precision parameter.}

\item{betas}{vector of beta coefficients for log-linear sampling model.}

\item{params}{output from initialize_parameters().}
}
\value{
numeric log-prior to agree with the output of mcmc_sampling(alg='ESS', samp_alg='MH').
}
\description{
Compute the log-prior using the set-up for ESS (sampling aware)
}
