% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_makers.R
\name{make_alignment}
\alias{make_alignment}
\title{Make alignment BEAST XML block}
\usage{
make_alignment(nexus_data)
}
\arguments{
\item{nexus_data}{character filename of nexus file containing the sequence 
data to be processed.}
}
\value{
a character string to be pasted into the alignment block of the BEAST
  XML.
}
\description{
Make alignment BEAST XML block
}

