% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_makers.R
\name{make_taxa}
\alias{make_taxa}
\title{Make taxa BEAST XML block}
\usage{
make_taxa(tree, units = "years")
}
\arguments{
\item{tree}{a phylo tree to be processed.}

\item{units}{character units of time for BEAST.}
}
\value{
a character string to be pasted into the taxa block of the BEAST XML.
}
\description{
Make taxa BEAST XML block
}

