% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phyloinfer.R
\name{mcmc_sampling}
\alias{mcmc_sampling}
\title{MCMC Sampling}
\usage{
mcmc_sampling(dataset, alg, nsamp, nburnin = 0, nsubsamp = 1, ngrid = 100,
  nugget = "1,1", prec_alpha = 0.01, prec_beta = 0.01, TrjL = NULL,
  Nleap = NULL, szkappa = NULL, rand_leap = NULL, f_init = rep(1, ngrid
  - 1), kappa = 1, covariates = NULL, betas = rep(0, 2 +
  length(covariates)), samp_alg = "none", kappa_alg = "gibbs",
  beta_vars = rep(100, length(betas)), printevery = 100)
}
\arguments{
\item{alg}{string selecting which MCMC sampler to use. Options are "HMC", 
"splitHMC", "MALA", "aMALA", and "ESS".}

\item{nsamp}{integer number of MCMC steps to compute.}

\item{nburnin}{integer number of MCMC steps to discard as burn-in.}

\item{nsubsamp}{integer after burn-in, how often to record a step to the 
output.}

\item{ngrid}{integer number of grid point in the latent field.}

\item{nugget}{string selecting which "nugget" adjustment to apply to the
precision matrix to make it full-rank. Options are '1,1' for an adjustment
to the first element, 'diag' for an adjustment to the entire main diagonal,
or 'none' which may result in a non-full-rank precision matrix.}

\item{prec_alpha}{numeric}

\item{prec_beta}{numeric}

\item{printevery}{}

\item{data}{\code{phylo} object or list containing vectors of coalescent 
times \code{coal_times}, sampling times \code{samp_times}, and number 
sampled per sampling time \code{n_sampled}.}
}
\description{
MCMC Sampling
}

