% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{mesa_traj}
\alias{mesa_traj}
\title{Mesa trajectory}
\usage{
mesa_traj(t, a = 2, b = 3)
}
\arguments{
\item{t}{numeric vector of times at which to evaluate the trajectory 
function.}

\item{a}{numeric radius of the mesa.}

\item{b}{numeric exponential slope of the mesa "walls".}
}
\value{
A vector containing the trajectory value corresponding to every 
  element in \code{t}.
}
\description{
Mesa trajectory
}
\examples{
mesa_traj(0:20)
}

