% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounded_coal_functions.R
\name{MLE_BoundCoal}
\alias{MLE_BoundCoal}
\title{Compute the MLE of effective population size
under the bounded coalescent}
\usage{
MLE_BoundCoal(data, lengthout = 5, bound, eps = 0.02, eta = 0.01)
}
\arguments{
\item{data}{A list of coal_times, samp_times and n_sampled
bound Value of upper bound on TMRCA}
}
\value{
grid and Ne estimates
}
\description{
Compute the MLE of effective population size
under the bounded coalescent
}
