% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NY_flu}
\alias{NY_flu}
\title{New York Influenza Genealogy}
\format{An object of class \code{phylo} of length 4.}
\usage{
NY_flu
}
\description{
This dataset is a \code{phylo} object of a genealogy generated by BEAST from
the sequence data in Rambaut, et al. 2008.
}
\details{
A. Rambaut, O. G. Pybus, M. I. Nelson, C. Viboud, J. K. Taubenberger, E. C. 
Holmes The genomic and epidemiological dynamics of human influenza A virus. 
Nature, 453(7195): 615-619, 2008.
}
\keyword{datasets}

