\name{phylodyn-package}
\alias{phylodyn-package}
\alias{phylodyn}
\docType{package}
\title{Tools for statistical genetics}
\description{
Tools for analyzing genome sequence data using INLA and MCMC.
}
\details{
\tabular{ll}{
Package: \tab phylodyn\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-05-24\cr
License: \tab What license is it under?\cr
}
##~~ An overview of how to use the package, including the most ~~
##~~ important functions ~~
}
\author{
Michael Karcher
Vladimir Minin

Maintainer: Michael Karcher <mkarcher@uw.edu>
}
\references{
##~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
##~~ Optional links to other man pages, e.g. ~~
##~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
##~~ simple examples of the most important functions ~~
}
