% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{plot_BEAST}
\alias{plot_BEAST}
\title{Plot trajectory estimate and credible intervals}
\usage{
plot_BEAST(midpts, med, up, low, cutoff = NULL, ylim = NULL,
  xlab = "Time", ylab = "Effective Population Size", axlabs = NULL,
  samp_times = NULL, n_sampled = NULL, nbreaks = 40, med_col = rgb(0.33,
  0.484, 0.828), cred_col = rgb(0.33, 0.484, 0.828, 0.4), heatmap_width = 7,
  heatmap_legend_cex = 1, ...)
}
\arguments{
\item{midpts}{numeric vector of x-values at which to plot estimate and credible interval bounds.}

\item{med}{numeric vector of estimated trajectory values.}

\item{up}{numeric vector of trajectory pointwise credible interval upper bounds.}

\item{low}{numeric vector of trajectory pointwise credible interval lower bounds.}

\item{cutoff}{numeric maximum time (in past) plotted.}

\item{ylim}{numeric y-axis interval.}

\item{xlab}{character x-axis label.}

\item{ylab}{character y-axis label.}

\item{axlabs}{character vector x-axis labels.}

\item{samp_times}{numeric vector of sampling times.}

\item{n_sampled}{integer vector of number of samples taken at each element of samp_times.}

\item{nbreaks}{integer number of bins in the sampling time histogram heatmap.}

\item{med_col}{color of trajectory estimate line.}

\item{cred_col}{color of trajectory pointwise credible interval region.}

\item{heatmap_width}{numeric vertical width of sampling time histogram heatmap.}

\item{heatmap_legend_cex}{numeric plotting expansion for heatmap legend.}

\item{...}{additional parameters to be passed to plot().}
}
\description{
Plot trajectory estimate and credible intervals
}

