% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{plot_seasonality}
\alias{plot_seasonality}
\title{Plot Seasonality}
\usage{
plot_seasonality(BNPR_out, zero_date, start = 0, years = NULL, period = 1,
  ylim = NULL, nbreaks = 40, lty = 1, lwd = 2, col_years = "gray",
  col_mean = "black", main = "", axlabs = NULL, xlab = "Time",
  xmarline = 3, ylab = "Effective Population Size", log_y = TRUE,
  heatmaps = TRUE, heatmap_labels = TRUE, heatmap_labels_side = "right",
  heatmap_width = 7, legend = NULL, yscale = 1, ...)
}
\arguments{
\item{BNPR_out}{output of BNPR or BNPR_PS.}

\item{zero_date}{}

\item{start}{}

\item{years}{}

\item{period}{}

\item{ylim}{}

\item{nbreaks}{}

\item{lty}{}

\item{lwd}{}

\item{col_years}{}

\item{col_mean}{}

\item{main}{}

\item{axlabs}{}

\item{xlab}{}

\item{xmarline}{}

\item{ylab}{}

\item{log_y}{}

\item{heatmaps}{}

\item{heatmap_labels}{}

\item{heatmap_labels_side}{}

\item{heatmap_width}{}

\item{legend}{}

\item{yscale}{}

\item{...}{}
}
\description{
Plot Seasonality
}

