% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_resolution.R
\name{posterior_beta_coal_hybrid}
\alias{posterior_beta_coal_hybrid}
\title{Joint inference of alpha that characterizes the Beta-coalescent and 
the effective population size trajectory via hybrid method}
\usage{
posterior_beta_coal_hybrid(tree, tol = 0.001, lengthout = 100, maxiter = 50)
}
\arguments{
\item{tree}{A \code{phylo} object.}

\item{tol}{tolerance for stopping the iteration.}

\item{lengthout}{numeric specifying number of grid points.}

\item{maxiter}{maximum number of iterations before termination if not converged.}
}
\value{
\code{alpha_opt} optimal \eqn{\alpha} value found by hybrid method, 
  \code{alpha_block_size_MLE} block-size MLE of \eqn{\alpha}, \code{niter}
  the number of iterations it took to converge, and \code{BNPR} the output of 
  BNPR using the \code{alpha_opt} value.
}
\description{
Joint inference of alpha that characterizes the Beta-coalescent and 
the effective population size trajectory via hybrid method
}
