% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_resolution.R
\name{posterior_beta_coal_MCMC_w_sHMC}
\alias{posterior_beta_coal_MCMC_w_sHMC}
\title{Joint inference of alpha that characterizes the Beta-coalescent and 
the effective population size trajectory via MCMC}
\usage{
posterior_beta_coal_MCMC_w_sHMC(
  tree,
  nsamp_total = 20000,
  lengthout = 100,
  burn_in = 0.1
)
}
\arguments{
\item{nsamp_total}{Integer representing the number of MCMC samples desired}

\item{lengthout}{numeric specifying number of grid points.}

\item{dataset}{\code{phylo} object or list containing vectors of coalescent 
times \code{coal_times}, sampling times \code{samp_times}, block sizes
\code{block_sizes}, and number sampled per sampling time \code{n_sampled}.}

\item{burn-in}{numeric between 0 and 1 indicating proportion of samples to toss 
at the start of the MCMC chain}
}
\value{
MCMC results of alpha and EPS sampled via sHMC: \code{\alpha} contains the 
  posterior sample of \eqn{\alpha},  \code{effpop} contains a vector of the posterior median 
  effective population size estimates, \code{grid} contains the grid end points, 
  \code{x} contains the grid point centers, \code{effpop025} and \code{effpop975} contain the 
  2.5th and 97.5th posterior percentiles, as well as the original list containing coalescent times,
  sampling times, block sizes, and number sampled per sampling time
}
\description{
Joint inference of alpha that characterizes the Beta-coalescent and 
the effective population size trajectory via MCMC
}
