% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_check.R
\name{posterior_coal_check}
\alias{posterior_coal_check}
\title{Posterior predictive check for coalescent model}
\usage{
posterior_coal_check(coal_times, samp_times, n_sampled, logpop, grid,
  cap = NULL)
}
\arguments{
\item{coal_times}{numeric vector of coalescent times.}

\item{samp_times}{numeric vector of sampling times.}

\item{n_sampled}{integer vector of lineages sampled at each sampling time.}

\item{logpop}{matrix posterior sample of log-effective population sizes.}

\item{grid}{numeric grid}

\item{cap}{integer maximum number of posterior samples to use.}
}
\description{
Posterior predictive check for coalescent model
}

