% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_check.R
\name{posterior_samp_check}
\alias{posterior_samp_check}
\title{Posterior predictive check for inhomogenous Poisson process sampling model}
\usage{
posterior_samp_check(samp_times, n_sampled, logpop, grid, betas,
  cap = NULL, covariates = NULL, cov_betas = NULL,
  pow_covariates = NULL, pow_cov_betas = NULL)
}
\arguments{
\item{samp_times}{numeric vector of sampling times.}

\item{n_sampled}{integer vector of lineages sampled at each sampling time.}

\item{logpop}{matrix posterior sample of log-effective population sizes.}

\item{grid}{numeric grid of time points.}

\item{betas}{matrix posterior sample of log-linear coefficients for sampling
model.}

\item{cap}{integer maximum number of posterior samples to use.}

\item{covariates}{list of vectors representing covariate values.}

\item{cov_betas}{numeric coefficients for each element of covariates.}

\item{pow_covariates}{list of vectors representing interaction covariate
values.}

\item{pow_cov_betas}{numeric coefficients for each element of pow_covariates.}
}
\value{
data.frame with two columns: "obs" contains the chi-square
  discrepancies for the supplied sampling times, "rep" contains the
  chi-square discrepancies for the replicated sampling times.
}
\description{
Posterior predictive check for inhomogenous Poisson process sampling model
}
