% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_resolution.R
\name{rbeta_coal}
\alias{rbeta_coal}
\title{Simulate a multifurcating tree under the Beta-coalescent}
\usage{
rbeta_coal(N, alpha, tip.label = NULL)
}
\arguments{
\item{N}{An integer giving the number of tips in the tree}

\item{alpha}{A number between 0 and 2 giving the parameter for the Beta coalescent}

\item{tip.label}{a character vector giving the tip labels; if not specified, the tips \code{t1, t2}, ..., are
given}
}
\value{
An object of class \code{phylo}
}
\description{
Simulate a multifurcating tree under the Beta-coalescent
}
