% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coal_lik_BSMC.R
\name{read_times}
\alias{read_times}
\title{Extracts coalescent times from a Multi-Phylo object}
\usage{
read_times(MyTree, sim, factor)
}
\arguments{
\item{MyTree}{\code{multiPhylo} object containing coalescent trees}

\item{sim}{integer specifying number of trees to read}

\item{factor}{numeric value to scale all the times}
}
\value{
A matrix of sim rows. Entry x_{i,j} has the n-j+1-th coalescent time of the i-th tree
}
\description{
Extracts coalescent times from a Multi-Phylo object
}
\examples{
read_times(ape::rmtree(3,5),3,1)
}

