% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SLT.R
\name{reformat_slt}
\alias{reformat_slt}
\title{Reformats data from 10x to be used with our method}
\usage{
reformat_slt(data, S)
}
\arguments{
\item{data}{a table with the following columns: cellBC, intBC, lineageGrp, Allele,r1,r2,..}

\item{S}{number of target sites 
Although the column Allele is not used, the function assumes target information starts in column 5.}
}
\value{
Dlist a list of cells x targets of alleles. One list per intBC
}
\description{
Reformats data from 10x to be used with our method
}
