% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{regional_flu}
\alias{regional_flu}
\title{Regional Influenza Genealogy}
\format{An object of class \code{list} of length 9.}
\usage{
regional_flu
}
\description{
This dataset is a \code{phylo} object of a genealogy generated by BEAST from
the sequence data in Zinder, et al. 2014.
}
\details{
D. Zinder, T. Bedford, E. B. Baskerville, R. J. Woods, M. Roy, M. Pascual.
Seasonality in the migration and establishment of H3N2 Influenza lineages
with epidemic growth and decline. BMC Evolutionary Biology, 14(1): 272, 2014.
}
\keyword{datasets}

