% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tajima_inference.R
\name{sample_genealogy}
\alias{sample_genealogy}
\title{Sample heterochronous topology.}
\usage{
sample_genealogy(SimTimeHet)
}
\arguments{
\item{SimTimeHet}{is the output of coalsim}
}
\value{
heterochronous genealogy
}
\description{
Sample heterochronous topology.
}
