% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalsieve.R
\name{sample_tree}
\alias{sample_tree}
\title{Sample a phylo tree}
\usage{
sample_tree(gene)
}
\arguments{
\item{gene}{a list containing}
}
\value{
a phylo tree consisent with `gene`.
}
\description{
Sample a phylo tree
}
\examples{
gene = coalsim(samp_times = c(0,1,2), n_sampled = c(2,1,1), traj = unif_traj)
tree = sample_tree(gene)
plot(tree)
}

