% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tajima_inference.R
\name{simulate_data_het}
\alias{simulate_data_het}
\title{Sample and superimpose mutations on a heterochronous genealogy}
\usage{
simulate_data_het(mu, tree, SimTimeHet)
}
\arguments{
\item{mu}{mutation rate}

\item{tree}{output of sample_genealogy}

\item{SimTimeHet}{output of coalsim}
}
\value{
sequence data in an incidence matrix format(rows are segregating sites, columns are samples)
}
\description{
Sample and superimpose mutations on a heterochronous genealogy
}
