% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_SLT.R
\name{simulate_SLData}
\alias{simulate_SLData}
\title{Simulation of alleles at S target sites evolving along the tree with M types}
\usage{
simulate_SLData(numSim = 1, tree, S = 2, M = 4, theta = diag(1, S))
}
\arguments{
\item{numSim}{number of simulated datasets}

\item{tree}{genealogical tree in ape format}

\item{S}{number of target sites}

\item{M}{number of different alleles}

\item{theta}{a matrix of mutation rates
It returns a list where the first
entry is the matrix of mutation states for the nodes, second entry is matrix of allele states.}
}
\description{
This function simulates alleles for all nodes according to a CTMC
}
