% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function countingv8.R
\name{siscount_tree}
\alias{siscount_tree}
\title{Compute Kingman perfect phylogeny}
\usage{
siscount_tree(n, mu, N, oldsuff)
}
\arguments{
\item{n}{sample size}

\item{mu}{mutation parameter}

\item{N}{SIS sample size}

\item{oldsuff}{Tajima perfect phylogeny}
}
\value{
estimates for the four resolutions and diagnostics.
}
\description{
Compute Kingman perfect phylogeny
}
