% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{sloped_traj}
\alias{sloped_traj}
\title{Sloped trajectory}
\usage{
sloped_traj(t)
}
\arguments{
\item{t}{numeric vector of times at which to evaluate the trajectory 
function.}
}
\value{
A vector containing the trajectory value corresponding to every 
  element in \code{t}.
}
\description{
Sloped trajectory
}
\examples{
sloped_traj(0:20)
}

