% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{steep_cyc_traj}
\alias{steep_cyc_traj}
\title{Steep cyclical trajectory}
\usage{
steep_cyc_traj(t)
}
\arguments{
\item{t}{numeric vector of times at which to evaluate the trajectory 
function.}
}
\value{
A vector containing the trajectory value corresponding to every 
  element in \code{t}.
}
\description{
Steep cyclical trajectory
}
\examples{
steep_cyc_traj(0:20)
}

