% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phyloinfer.R, R/tajima_inference.R
\name{subsample}
\alias{subsample}
\title{Subsample MCMC}
\usage{
subsample(res, burnin = 0, subsample = 1)

subsample(res, burnin = 0, subsample = 1)
}
\arguments{
\item{res}{realization of the chain}

\item{burnin}{burn of the chain}

\item{subsample}{subsampling of the MC chain}
}
\value{
A list with initial values.

approximate sample from the posterior
}
\description{
Subsample MCMC

Subsample
}
