% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tajima.R
\name{sufficient_stats}
\alias{sufficient_stats}
\title{Compute Gusfield Tajima perfect phylogeny}
\usage{
sufficient_stats(data)
}
\arguments{
\item{data}{in the format of the function simulate_data}
}
\value{
PP Tajima
}
\description{
Compute Gusfield Tajima perfect phylogeny
}
