% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_resolution.R
\name{summarize_multif_phylo}
\alias{summarize_multif_phylo}
\title{Summarize a multifurcating phylogeny.}
\usage{
summarize_multif_phylo(tree)
}
\arguments{
\item{phy}{a \code{phylo} object containing a phylogeny.}
}
\value{
A list containing vectors of sampling times \code{samp_times}, number 
  sampled per sampling time \code{n_sampled}, coalescent times \code{coal_times},
  and block sizes per coalescent event \code{block_sizes}.
}
\description{
Summarize a multifurcating phylogeny.
}
