% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_newick.R
\name{summarize_phylo}
\alias{summarize_phylo}
\title{Summarize a phylogeny.}
\usage{
summarize_phylo(phy)
}
\arguments{
\item{phy}{a \code{phylo} object containing a phylogeny.}
}
\value{
A list containing vectors of sampling times \code{samp_times}, number 
  sampled per sampling time \code{n_sampled}, and coalescent times
  \code{coal_times}.
}
\description{
Summarize a phylogeny.
}
\examples{
data("NY_flu")
summarize_phylo(NY_flu)
}

