% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tajima_inference.R
\name{tajima_perphylo}
\alias{tajima_perphylo}
\title{Prepare the perfect phylogeny to be used for plotting (not the one used in the actual code)}
\usage{
tajima_perphylo(data)
}
\arguments{
\item{data}{in the Tajima format (rows are segrating sites and columns are the sequence)}
}
\value{
perfect phylogeny
}
\description{
Prepare the perfect phylogeny to be used for plotting (not the one used in the actual code)
}
