% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{unif_traj}
\alias{unif_traj}
\title{Uniform trajectory.}
\usage{
unif_traj(t, level = 100)
}
\arguments{
\item{t}{numeric vector of times at which to evaluate the trajectory 
function.}

\item{level}{numeric. The value to return at all values \code{t}.}
}
\value{
A vector containing a value of \code{level} for every element in
  \code{t}.
}
\description{
Uniform trajectory.
}
\examples{
unif_traj(0:10, level=5)
}

