% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phyloinfer.R
\name{updateFmat_Markov_FREE}
\alias{updateFmat_Markov_FREE}
\title{HMC for times}
\usage{
updateFmat_Markov_FREE(
  currentval,
  F_list,
  probs_list,
  const = 1,
  nsites = nsites,
  oldsuff,
  n
)
}
\arguments{
\item{F_list}{currentvalue F matrix}

\item{probs_list}{matrix with likelihod posterior prior at each iteration}

\item{const}{scaling factor}

\item{nsites}{mutation rate}

\item{oldsuff}{perfect phylogeny}

\item{n}{sample size}

\item{currenval}{current state of the chain}
}
\value{
A list with initial values.
}
\description{
HMC for times
}
