% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tajima_inference.R
\name{updateFmat_Markov_FREE_het}
\alias{updateFmat_Markov_FREE_het}
\title{Moves topology Cappello et al. (2020)}
\usage{
updateFmat_Markov_FREE_het(
  currentval,
  probs_list,
  const = 1,
  oldsuff,
  n,
  type = "Tajima"
)
}
\arguments{
\item{currentval}{current val of the chain}

\item{probs_list}{array recording probability}

\item{const}{scaling factor}

\item{oldsuff}{perfect phylogeny}

\item{n}{vector of samples}

\item{mu}{mutation parameter}
}
\value{
MC steps for topology
}
\description{
Moves topology Cappello et al. (2020)
}
