% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phyloinfer.R
\name{updateSingleTime}
\alias{updateSingleTime}
\title{Local updates for times}
\usage{
updateSingleTime(
  currentval,
  theta_list,
  prec_list,
  probs_list,
  const = 1,
  nsites = nsites
)
}
\arguments{
\item{theta_list}{value of log Ne(t)}

\item{prec_list}{vector recording precision parameter}

\item{probs_list}{matrix with likelihod posterior prior at each iteration}

\item{const}{scaling factor}

\item{nsites}{mutation rate}

\item{currenval}{current state of the chain}
}
\value{
A list with initial values.
}
\description{
Local updates for times
}
