% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phyloinfer.R
\name{updateTheta}
\alias{updateTheta}
\title{HMC for theta}
\usage{
updateTheta(
  currentval,
  theta_list,
  prec_list,
  probs_list,
  const = 1,
  j = 1,
  nsites = nsites,
  eps = 0.3
)
}
\arguments{
\item{theta_list}{all value of log Ne(t) (record all iterations)}

\item{prec_list}{vector recording precision parameter}

\item{probs_list}{matrix with likelihod posterior prior at each iteration}

\item{const}{scaling factor}

\item{j}{decide whether to use gradient or likehood}

\item{nsites}{mutation rate}

\item{eps}{leapfrof parameter}

\item{currenval}{current state of the chain}
}
\value{
A list with initial values. theta is log Ne
}
\description{
HMC for theta
}
