% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tajima_inference.R
\name{updateTheta_het}
\alias{updateTheta_het}
\title{HMC step}
\usage{
updateTheta_het(
  currentval,
  theta_list,
  prec_list,
  probs_list,
  const = 1,
  j = 1,
  nsites = nsites,
  eps = 0.3
)
}
\arguments{
\item{currentval}{current val of the chain}

\item{theta_list}{array in which each column a realization of the chain.}

\item{prec_list}{vector recording precision parameters}

\item{probs_list}{array recording probability}

\item{const}{scaling factor}

\item{j}{}

\item{nsites}{mutation parameter}

\item{eps}{step size}
}
\value{
MC steps for effective population size
}
\description{
HMC step
}
