% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tajima_inference.R
\name{updateTimes_hetLocal}
\alias{updateTimes_hetLocal}
\title{Global Local step for times}
\usage{
updateTimes_hetLocal(
  n,
  oldsuff,
  currentval,
  times_list,
  theta_list,
  prec_list,
  probs_list,
  const = 1,
  mu,
  samp_times,
  a = 2,
  sigma = 0.2
)
}
\arguments{
\item{n}{vector of samples size}

\item{oldsuff}{perfect phylogeny}

\item{currentval}{current val of the chain}

\item{times_list}{array in which each column a realization of the chain.}

\item{theta_list}{array in which each column a realization of the chain.}

\item{prec_list}{vector recording precision parameters}

\item{probs_list}{array recording probability}

\item{const}{scaling factor}

\item{mu}{mutation parameter}

\item{samp_times}{vector of sampling times}

\item{a}{Z1}

\item{sigma}{sdev of the truncated normal.}
}
\value{
MC steps for coalescent times
}
\description{
Global Local step for times
}
