% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{BNPR_samp_only}
\alias{BNPR_samp_only}
\title{Nonparametric estimate of sampling intensity}
\usage{
BNPR_samp_only(data, lengthout = 100, grid = NULL, tmrca = NULL,
  prec_alpha = 0.01, prec_beta = 0.01)
}
\arguments{
\item{data}{\code{phylo} object or list containing vectors of sampling times
\code{samp_times} and number sampled per sampling time \code{n_sampled}.}

\item{lengthout}{numeric specifying number of grid points.}

\item{grid}{numeric vector of endpoints of the latent field.}

\item{prec_alpha, prec_beta}{numerics specifying gamma prior for precision 
\eqn{\kappa}.}
}
\description{
Nonparametric estimate of sampling intensity
}

