% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coalsieve.R
\name{coalsim}
\alias{coalsim}
\title{Simulate from inhomogeneous, heterochronous coalescent.}
\usage{
coalsim(s_times, n_sampled, traj, upper = 25, ...)
}
\arguments{
\item{s_times}{numeric vector of sampling times.}

\item{n_sampled}{numeric vector of samples taken per sampling time.}

\item{traj}{function that returns effective population size at time t.}

\item{upper}{numeric upper limit on \code{traj} function on its support.}

\item{...}{additional arguments to be passed to \code{traj} function.}
}
\value{
A list containing vectors of coalescent times \code{coal_times},
  intercoalescent times \code{intercoal_times}, and number of active lineages
  \code{lineages}.
}
\description{
Simulate from inhomogeneous, heterochronous coalescent.
}
\examples{
coalsim(0:2, 3:1, unif_traj, upper=10, level=10)
}

