% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{plot_mrw}
\alias{plot_mrw}
\title{Plot Mean Relative Widths}
\usage{
plot_mrw(BNPR_outs, traj = NULL, xlim = NULL, nbreaks = 40, ltys = 1,
  lwds = 2, cols = "black", xlab = "Time", xmarline = 3,
  axlabs = NULL, ylim = NULL, ymin_zero = FALSE,
  ylab = "Mean Relative Width", main = "", heatmaps = TRUE,
  heatmap_labels = FALSE, heatmap_labels_side = "right",
  heatmap_width = 7, legends = NULL, legend_place = "topleft",
  legend_cex = 1, bty = "n", ...)
}
\arguments{
\item{BNPR_outs}{one or more outputs of BNPR or BNPR_PS.}

\item{traj}{function summarizing the true effective population size 
trajectory.}

\item{xlim}{}

\item{nbreaks}{}

\item{ltys}{}

\item{lwds}{}

\item{cols}{}

\item{xlab}{}

\item{xmarline}{}

\item{axlabs}{}

\item{ylim}{}

\item{ymin_zero}{}

\item{ylab}{}

\item{main}{}

\item{heatmaps}{}

\item{heatmap_labels}{}

\item{heatmap_labels_side}{}

\item{heatmap_width}{}

\item{legends}{}

\item{legend_place}{}

\item{legend_cex}{}

\item{bty}{}

\item{...}{}
}
\description{
Plot Mean Relative Widths
}

