% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_check.R
\name{posterior_samp_check}
\alias{posterior_samp_check}
\title{Posterior predictive check for inhomogenous Poisson process sampling model}
\usage{
posterior_samp_check(samp_times, n_sampled, logpop, grid, betas, cap = NULL)
}
\arguments{
\item{samp_times}{numeric vector of sampling times.}

\item{n_sampled}{integer vector of lineages sampled at each sampling time.}

\item{logpop}{matrix posterior sample of log-effective population sizes.}

\item{grid}{numeric grid}

\item{betas}{matrix posterior sample of log-linear coefficients for sampling
model.}

\item{cap}{integer maximum number of posterior samples to use.}
}
\description{
Posterior predictive check for inhomogenous Poisson process sampling model
}

